Big Bugfix: rsPow has been renamed to rsPowI, etc.
We had a problem with overload resolution. Sometimes a function called rsPow was called that defined a templatized first parameter and integer second parameter - but when it was called with two doubles, then the second argument got converted to integer - which was totally not intended and produced wrong results in such cases! That function has now been renamed to rsPowI and another function template rsPow was added that defined two template parameters of the same type. Now this gets called, as intended, when instantiating the template with two doubles, for example.