rsSparseFilter::updateDelayLineLength() doesn't re-allocate anymore
Previously, it just increased the capacity of the delayline if needed. Now it asserts that the delayline already has enough capacity. This is because we want to make this function realtime safe. Now the onus is on client code to pre-allocate enough delay memory for the filter in some non-realtime function (probably on construction or on some "prepareToPlay" function)