Recreate IDE project files for ToolChain with Projucer v 8.0.8
Formerly, I have used Projucer 6.0.8 to generate the IDE project files. After the update to the new version 8.0.8, Projucer now doesn't support to export to VS 2015 anymore, so the project exporter for VS 2015 is now gone from the ToolChain.jucer file. On the other hand, there's now a new exporter for VS 2022. I also updated the .gitignore file to let it ignore a couple of files and subdirectories from the new VisualStudio2022 folder. Apparently, the update to the new Projucer also caused some changes in the generated project files for XCode and the Linux makefile.