Move rsTrunc(), rsFloor(), rsCeil() into Fraction.h
These functions were previously located in a preliminary location within the unit tests. But we need them now in the library proper to use them in the context of some experiments with continued fractions. I'm not sure, if they are already production ready, though. They may still be suboptimal efficiency-wise. But they do pass their unit tests, so they should be ok to use.