rsModalFilterWithAttack allows now for an attack of zero
The general coefficient computation formulas will have a division by zero when the user requests an attack time of zero. Therefore, we handle this case now as a special case. But this is a preliminary solution. The real solution will be to make sure that the design formulas themselves can handle this case.