combVsModalBank() now uses delay as user parameter
Previously, the user had to set up M, the number of modes and the delay was then calculated as 2*M. This implied that we could only use even delay lengths. Now we allow also for setting up odd delays - but it doesn't work correctly yet with odd delays - that will be fixed in the next commit, hopefully.